README.md
Countdown to Christmas
This program is a simple countdown to Christmas in Python. It calculates the difference between today's date and Christmas, and then prints the number of days in between.

Copyright (c) 2023 Fox Studios - Scratch™

This software is licensed under the MIT License.

Usage:

python
python countdown_to_christmas.py

Output:

There are {countdown.days} days until Christmas!
Example:

python countdown_to_christmas.py
There are 3 days until Christmas!
Customization:

You can customize this program to fit your needs, such as changing the format of the output or adding additional features. For example, you could add the ability to set a custom date for the countdown, or you could add a notification system that alerts you when Christmas is approaching.

Contributions:

If you have any suggestions for how to improve this program, please feel free to contribute to the project. 

License:

This program is licensed under the MIT License.

You are free to use, modify, and distribute this software for any purpose, with or without credit to the author.